When the graph data is scoped by graph.window.{xMin, xMax} add points on the xMin and yMax extremes of the chart representing the first datapoints before and after the range.
The y-value of the new datapoint at xMin is chosen so that it lies on a
line between the first datapoint before the range and the first
datapoint in the range. The y-value of the new datapoint at xMax is
chosen similarly.

A flag 'preventHover' is set on the new datapoints so that
the HoverDetail renderer will not add hover hanling for these points.