Use the graph's window.{xMin,xMax} to set the range of the xaxis for a chart and the preview slider.
If you increase the preview selection to include a region that contains
no new datapoints then the main chart will show a corresponding blank
region.

Previously you had to include new datapoints in the preview
selection for the main chart to update. This meant that the selected
region in the preview and the main chart were not synchronized.

(depends on
https://github.com/paulcadman/rickshaw/commit/6fd70b9b203551ae7236a2a6e6b5d74bf2cba656).
https://github.com/shutterstock/rickshaw/issues/345